app/tg-bot/requirements.txt
numpy
Pillow
asyncio
aiogram==3.7.0
python-dotenv
requests